+++ CHANGELOG.md
@@ -1,4 +1,8 @@
-## 1.0.2
+## 1.1.0
+
+* Add `StreamChannel.transformStream()`, `StreamChannel.transformSink()`,
+  `StreamChannel.changeStream()`, and `StreamChannel.changeSink()` to support
+  changing only the stream or only the sink of a channel.
 
 * Be more explicit about `JsonDocumentTransformer`'s error-handling behavior.